EGL/Wayland: DRI_PRIME support v3
Implements DRI_PRIME support for the Wayand EGL backend.

v3: Add a check to dri2_create_image_wayland_wl_buffer to fail
if we forbid to use tiling (which now means we use another gpu
than the compositor). We can't check the buffer is untiled (which
is unlikely), so we fail here. It's in the specification.
v2: some cleanup and fixes

Signed-off-by: Axel Davy <axel.davy@ens.fr>