EGL/Wayland drirc: Use drirc to complement DRI_PRIME
Add to drirc a new parameter: "wanted_device_id_path_tag"
to indicate the gpu we would like to use.

The EGL Wayland platform checks this parameter under
the driver name "init". It should be possible to get GLX
to check this parameter too, and use it for DRI3
GPU offlading.

Signed-off-by: Axel Davy <axel.davy@ens.fr>