Wayland/egl: Add GPU offloading support
This is easier than GLX DRI3 GPU offloading support,
because applications are not allowed to read the front
buffer.

We just need to send to the server buffers it can read,
and for that we use for every back buffer an intermediate
buffer with no tiling to which we copy before sending
the buffer to the compositor.

Signed-off-by: Axel Davy <axel.davy@ens.fr>