GLX: Add an env var to enable the support of GLX extensions needing both client and server support, in case of client support and direct rendering.
In the case of XWayland, there's no accelerated indirect rendering.
For example GLX_ARB_create_context is not advertised by the server,
and according to the spec, we are not allowed to advertise it
to the application because of that.

This env var makes Mesa ignore this restriction, and
a GLX extension is advertised whenever the client supports it.

Signed-off-by: Axel Davy <axel.davy@ens.fr>